pull assessment csv loading out
